requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Caching==2.3.1
Flask-Migrate==4.1.0
WTForms==3.2.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
gunicorn==23.0.0
APScheduler==3.11.0
requests==2.32.3
beautifulsoup4==4.12.3
python-slugify==8.0.4
Werkzeug==3.1.3
email-validator==2.2.0
python-telegram-bot==21.9
trafilatura==2.0.0
pdfminer.six==20260107
PyPDF2==3.0.1
